requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
motor==3.7.1
pymongo==4.13.2
pydantic-settings==2.10.1
opencv-python==4.12.0.88
numpy==2.1.3
Pillow==11.3.0
python-dotenv==1.1.1
ultralytics==8.3.187
requests==2.32.5
